+++ cmd/logInProgress.go
@@ -15,12 +15,8 @@
 package cmd
 
 import (
-	"io"
-	"os"
-
 	"github.com/lucassabreu/clockify-cli/api"
 	"github.com/lucassabreu/clockify-cli/api/dto"
-	"github.com/lucassabreu/clockify-cli/internal/output"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
@@ -36,9 +32,7 @@ var logInProgressCmd = &cobra.Command{
 			return err
 		}
 
-		format, _ := cmd.Flags().GetString("format")
-		asJSON, _ := cmd.Flags().GetBool("json")
-		return formatTimeEntry(te, asJSON, format)
+		return formatTimeEntry(te, cmd)
 	}),
 }
 
@@ -55,30 +49,9 @@ func getTimeEntryInProgres(c *api.Client, workspace string) (*dto.TimeEntry, err
 	return &tef, err
 }
 
-func formatTimeEntry(te *dto.TimeEntry, asJSON bool, format string) error {
-	var reportFn func(*dto.TimeEntry, io.Writer) error
-
-	reportFn = output.TimeEntryPrint
-
-	if asJSON {
-		reportFn = output.TimeEntryJSONPrint
-	}
-
-	if format != "" {
-		reportFn = output.TimeEntryPrintWithTemplate(format)
-	}
-
-	return reportFn(te, os.Stdout)
-}
-
 func init() {
 	logCmd.AddCommand(logInProgressCmd)
-	setCommonFormats(logInProgressCmd)
+	addPrintTimeEntriesFlags(logInProgressCmd)
 
 	_ = logInProgressCmd.MarkFlagRequired(WORKSPACE)
-}
-
-func setCommonFormats(cmd *cobra.Command) {
-	cmd.Flags().StringP("format", "f", "", "golang text/template format to be applied")
-	cmd.Flags().BoolP("json", "j", false, "print as json")
 }